Separate sidebar for mobile (see description for reasons)
This remove the unnecessary complexity of deciding default sidebar state in SSR.
On wide screen, we want the sidebar to be opened by default, and vice versa on small screen.
The problem is that we can't access window in the server to get media query.
Deriving it from user agent is also complex and not reliable.